+++ lib/WP_Auth0_Api_Client.php
@@ -2,6 +2,68 @@
 
 class WP_Auth0_Api_Client {
 
+	private static $connect_info = null;
+
+	/**
+	 * Generate the API endpoint with a provided domain
+	 *
+	 * @since 3.5.0
+	 *
+	 * @param string $path - API path appended to the domain
+	 * @param string $domain - domain to use, blank uses default
+	 *
+	 * @return string
+	 */
+	private static function get_endpoint( $path = '', $domain = '' ) {
+
+		if ( empty( $domain ) ) {
+			$a0_options = WP_Auth0_Options::Instance();
+			$domain = $a0_options->get( 'domain' );
+		}
+
+		if ( ! empty( $path[0] ) && '/' === $path[0] ) {
+			$path = substr( $path, 1 );
+		}
+
+		return "https://{$domain}/{$path}";
+	}
+
+	/**
+	 * Return basic connection information, or a specific value
+	 *
+	 * @since 3.5.0
+	 *
+	 * @param string $opt - specific option needed, returns all if blank
+	 *
+	 * @return string|array
+	 */
+	public static function get_connect_info( $opt = '' ) {
+
+		if ( is_null( self::$connect_info ) ) {
+			$a0_options = WP_Auth0_Options::Instance();
+
+			self::$connect_info = array(
+				'domain' => $a0_options->get( 'domain' ),
+				'client_id' => $a0_options->get( 'client_id' ),
+				'client_secret' => $a0_options->get( 'client_secret' ),
+				'client_secret_encoded' => $a0_options->get( 'client_secret_b64_encoded' ),
+				'connection' => $a0_options->get( 'db_connection_name' ),
+				'app_token' => $a0_options->get( 'auth0_app_token' ),
+				'audience' => $a0_options->get( 'api_audience' ),
+			);
+
+			if ( empty( self::$connect_info[ 'audience' ] ) ) {
+				self::$connect_info[ 'audience' ] = self::get_endpoint( 'api/v2/' );
+			}
+		}
+
+		if ( empty( $opt ) ) {
+			return self::$connect_info;
+		} else {
+			return ! empty( self::$connect_info[ $opt ] ) ? self::$connect_info[ $opt ] : '';
+		}
+	}
+
 	public static function ro( $domain, $client_id, $username, $password, $connection, $scope ) {
 
 		$endpoint = "https://$domain/";
@@ -88,6 +150,31 @@ class WP_Auth0_Api_Client {
 		);
 	}
 
+	/**
+	 * Basic header components for an Auth0 API call
+	 *
+	 * @since 3.5.0
+	 *
+	 * @param string $token - for Authorization header
+	 * @param string $content_type - for Content-Type header
+	 *
+	 * @return array
+	 */
+	private static function get_headers( $token = '', $content_type = 'application/json' ) {
+
+		$headers = self::get_info_headers();
+
+		if ( ! empty( $token ) ) {
+			$headers['Authorization'] = "Bearer {$token}";
+		}
+
+		if ( ! empty( $content_type ) ) {
+			$headers[ 'Content-Type' ] = $content_type;
+		}
+
+		return $headers;
+	}
+
 	public static function get_token( $domain, $client_id, $client_secret, $grantType = 'client_credentials', $extraBody = null ) {
 		if ( ! is_array( $extraBody ) ) {
 			$body = array();
@@ -117,6 +204,42 @@ class WP_Auth0_Api_Client {
 		}
 
 		return $response;
+	}
+
+	/**
+	 * Get a client_credentials token using default stored connection info
+	 *
+	 * @since 3.4.1
+	 *
+	 * @return bool|string
+	 */
+	public static function get_client_token() {
+
+		$response = wp_remote_post( self::get_endpoint( 'oauth/token' ), array(
+				'headers' => self::get_headers(),
+				'body' => json_encode( array(
+					'client_id' => self::get_connect_info( 'client_id' ),
+					'client_secret' => self::get_connect_info( 'client_secret' ),
+					'audience' => self::get_connect_info( 'audience' ),
+					'grant_type' => 'client_credentials',
+				) ),
+			) );
+
+		if ( $response instanceof WP_Error ) {
+			WP_Auth0_ErrorManager::insert_auth0_error( __METHOD__, $response );
+			error_log( $response->get_error_message() );
+			return false;
+		}
+
+		if ( $response['response']['code'] !== 200 ) {
+			WP_Auth0_ErrorManager::insert_auth0_error( __METHOD__, $response['body'] );
+			error_log( $response['body'] );
+			return false;
+		}
+
+		$response = json_decode( $response['body'] );
+
+		return ! empty( $response->access_token ) ? $response->access_token : '';
 	}
 
 	public static function get_user_info( $domain, $access_token ) {
@@ -149,6 +272,43 @@ class WP_Auth0_Api_Client {
 		return json_decode( $response['body'] );
 	}
 
+	/**
+	 * Trigger a verification email re-send
+	 *
+	 * @since 3.5.0
+	 *
+	 * @param string $user_id - Auth0 formatted user_id
+	 *
+	 * @return bool
+	 */
+	public static function resend_verification_email( $user_id ) {
+
+		$response = wp_remote_post(
+			self::get_endpoint( 'api/v2/jobs/verification-email' ),
+			array(
+				'headers' => self::get_headers( self::get_client_token() ),
+				'body' => json_encode( array(
+					'user_id' => $user_id,
+					'client_id' => self::get_connect_info( 'client_id' ),
+				) ),
+			)
+		);
+
+		if ( $response instanceof WP_Error ) {
+			WP_Auth0_ErrorManager::insert_auth0_error( __METHOD__, $response );
+			error_log( $response->get_error_message() );
+			return false;
+		}
+
+		if ( $response['response']['code'] !== 201 ) {
+			WP_Auth0_ErrorManager::insert_auth0_error( __METHOD__, $response['body'] );
+			error_log( $response['body'] );
+			return false;
+		}
+
+		return true;
+	}
+
 	public static function get_user( $domain, $jwt, $user_id ) {
 		$endpoint = "https://$domain/api/v2/users/" . urlencode( $user_id );
 
@@ -159,6 +319,8 @@ class WP_Auth0_Api_Client {
 		return wp_remote_get( $endpoint  , array(
 				'headers' => $headers,
 			) );
+
+
 	}
 
 	public static function create_user( $domain, $jwt, $data ) {
@@ -226,6 +388,7 @@ class WP_Auth0_Api_Client {
 			'read:connections',
 			'create:rules',
 			'delete:rules',
+			'read:users',
 			'update:users',
 			'update:guardian_factors',
 		);
@@ -240,29 +403,25 @@ class WP_Auth0_Api_Client {
 		$headers['Authorization'] = "Bearer $app_token";
 		$headers['content-type'] = "application/json";
 
-		$logout_url = home_url();
-
 		$response = wp_remote_post( $endpoint  , array(
 				'method' => 'POST',
 				'headers' => $headers,
 				'body' => json_encode( array(
 						'name' => $name,
 						'callbacks' => array(
-							home_url( '/index.php?auth0=1' ),
-							home_url( '/wp-login.php' )
+							site_url( 'index.php?auth0=1' ),
+							wp_login_url()
 						),
 						"allowed_origins"=>array(
-							home_url( '/wp-login.php' )
+							wp_login_url()
 						),
 						"jwt_configuration" => array(
 							"alg" => "RS256"
 						),
 						"app_type" => "regular_web",
 						"cross_origin_auth" => true,
-						"cross_origin_loc" => home_url('/index.php?auth0fallback=1','https'),
-						"allowed_logout_urls" => array(
-							$logout_url
-						),
+						"cross_origin_loc" => site_url('index.php?auth0fallback=1','https'),
+						"allowed_logout_urls" => array( wp_logout_url() ),
 					) )
 			) );
 
@@ -282,11 +441,11 @@ class WP_Auth0_Api_Client {
 	
 		// Workaround: Can't add `web_origin` on create
 		$payload = array(
-			"web_origins" => array(home_url())
+			"web_origins" => ( home_url() === site_url() ? array( home_url() ) : array( home_url(), site_url() ) )
 		);
 		$updateResponse = WP_Auth0_Api_Client::update_client($domain, $app_token, $response->client_id, false, $payload);
 
-		if ( $updateClient instanceof WP_Error ) {
+		if ( $updateResponse instanceof WP_Error ) {
 			WP_Auth0_ErrorManager::insert_auth0_error( 'WP_Auth0_Api_Client::create_client', $updateResponse );
 			error_log( $updateResponse->get_error_message() );
 			return false;
@@ -411,6 +570,57 @@ class WP_Auth0_Api_Client {
 
 		if ( $response['response']['code'] != 204 ) {
 			WP_Auth0_ErrorManager::insert_auth0_error( 'WP_Auth0_Api_Client::delete_rule', $response['body'] );
+			error_log( $response['body'] );
+			return false;
+		}
+
+		return json_decode( $response['body'] );
+	}
+
+	/**
+	 * Create a client grant for the management API
+	 *
+	 * @param $app_token
+	 * @param $client_id
+	 *
+	 * @return array|bool|mixed|object
+	 */
+	public static function create_client_grant( $app_token, $client_id ) {
+
+		$data = array(
+			'client_id' => $client_id,
+			'audience' => self::get_connect_info( 'audience' ),
+			'scope' => self::get_required_scopes()
+		);
+
+		$response = wp_remote_post( self::get_endpoint( 'api/v2/client-grants' ), array(
+			'headers' => self::get_headers( $app_token ),
+			'body' => json_encode( $data ),
+		) );
+
+		if ( $response instanceof WP_Error ) {
+			WP_Auth0_ErrorManager::insert_auth0_error( __METHOD__, $response );
+			error_log( $response->get_error_message() );
+			return false;
+		}
+
+		if ( 409 === $response['response']['code'] ) {
+
+			// Client grant from WP-created client to Management API already exists
+			WP_Auth0_ErrorManager::insert_auth0_error(
+				__METHOD__,
+				sprintf(
+					__( 'A client grant for %s to %s has already been created. Make sure this grant at least includes %s.', 'wp-auth0' ),
+					self::get_connect_info( 'client_id' ),
+					self::get_connect_info( 'audience' ),
+					implode( ', ', self::get_required_scopes() )
+				)
+			);
+			return true;
+
+		} else if ( $response['response']['code'] != 201 ) {
+
+			WP_Auth0_ErrorManager::insert_auth0_error(  __METHOD__, $response['body'] );
 			error_log( $response['body'] );
 			return false;
 		}
@@ -704,6 +914,9 @@ class WP_Auth0_Api_Client {
 		return array(
 			'create:clients',
 			'update:clients',
+
+			'create:client_grants',
+			'update:client_grants',
 
 			'update:connections',
 			'create:connections',