+++ templates/auth0-login-form-lock10.php
@@ -80,18 +80,11 @@ document.addEventListener("DOMContentLoaded", function() {
             return p;
           }, {});
 
-          post('<?php echo home_url( '/index.php?auth0=implicit' ); ?>', {
+          post('<?php echo site_url( 'index.php?auth0=implicit' ); ?>', {
             token:data.id_token,
             state:data.state
           }, 'POST');
         }
-
-        // lock.on("authenticated", function(authResult) {
-        //   post('<?php echo home_url( '/index.php?auth0=implicit' ); ?>', {
-        //     token:authResult.idToken,
-        //     state:authResult.state
-        //   }, 'POST');
-        // });
 
         function post(path, params, method) {
             method = method || "post"; // Set method to post by default if not specified.
@@ -123,10 +116,7 @@ document.addEventListener("DOMContentLoaded", function() {
             document.body.appendChild(form);
             form.submit();
         }
-
-        // function a0ShowLoginModal() {
-        //     lock.<?php echo $lock_options->get_lock_show_method(); ?>();
-        // }
+    
     <?php } ?>
 
     if (!ignore_sso) {